+++ .formatter.exs
@@ -6,7 +6,9 @@ locals_without_parens = [
   assert_called: 4,
   refute_called: 2,
   refute_called: 3,
-  refute_called: 4
+  refute_called: 4,
+  assert_called!: 2,
+  assert_called!: 3
 ]
 
 [